requirements.txt
torch-directml
pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
tqdm>=4.50.0
tensorboard>=2.7.0
scikit-image>=0.18.0